fix shape tests
